SongController Bug Fixes

+++ trianafy-base-main/src/main/java/com/salesianostriana/dam/trianafy/controller/SongController.java
@@ -96,10 +96,9 @@ public class SongController {
 
     @DeleteMapping("/song/{id}")
     public ResponseEntity<?> deleteSong(@PathVariable Long id) {
-        songRepo.deleteById(id);
-        return ResponseEntity
-                .noContent()
-                .build();
+        if (songRepo.existsById(id))
+            songRepo.deleteById(id);
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 
 }
